feat: add background pattern
